requirements_openai.txt
# Requirements for End-to-End Prompt Generation (RL + OpenAI)
# Install with: pip install -r requirements_openai.txt

# Core RL dependencies
gymnasium>=0.29.1
stable-baselines3>=2.3.0
numpy>=1.26.0
tqdm>=4.66.0

# OpenAI integration
openai>=1.0.0

# Additional utilities
python-dotenv>=1.0.0  # For environment variable management
